refactor: split rdg columns generating code from MessagesTab.jsx
